pulls into standalone app module for better testing
